+++ src/defs.rs
@@ -27,8 +27,8 @@ impl NrOf {
     pub const PIECE_TYPES: usize = 6;
     pub const CASTLING_PERMISSIONS: usize = 16; // 0-15
     pub const SQUARES: usize = 64;
-    pub const FILES: u8 = 8;
-    pub const RANKS: u8 = 8;
+    pub const FILES: usize = 8;
+    pub const RANKS: usize = 8;
 }
 
 pub struct Castling;
